refactor inputBarButtons and merge files
